requirements.txt
pydantic
GitPython
PyGithub
python-dotenv
requests